Split _print into _print.json and _print.rich
