+++ src/pypi_changes/__init__.py
@@ -5,7 +5,8 @@ from typing import Sequence
 from ._cli import parse_cli_arguments
 from ._distributions import collect_distributions
 from ._info import pypi_info
-from ._print import print_json, print_tree
+from ._print.json import print_json
+from ._print.rich import print_tree
 from ._version import version
 
 #: semantic version of the package
